refactor: move whatsapp handler wiring into platform
